experimentation/requirements.txt
tqdm
freshstack
PyYAML>=6.0
traitlets
transformers
nbconvert
semchunk
tree-sitter
tree-sitter-language-pack
torch
tiktoken
ipython
datasets
sentence_transformers
pyserini
faiss-gpu-cu12
voyageai
numpy
matplotlib
scikit-learn
accelerate
